add player character, bat, weapon and hp Bar
i made the player movement and animation also i made the weapon and set the damage of the weapon and hp of the player. i made the hp bar which detect how much hp the player still have. i set the animation of the bat which is the first enemy and made it damage the player character when there is a collision